+++ src/lib.rs
@@ -46,6 +46,7 @@ pub use la_policy::{LAPolicy, Policy};
 pub use la_public_key::{LAPublicKey, SecKeyAlgorithm, SecKeyExchangeParameters};
 pub use la_right::{LARight, LARightState};
 pub use la_right_store::LARightStore;
+pub use security::{AccessControl, AccessControlFlags, AccessControlProtection};
 pub use timeout::{set_sync_timeout, sync_timeout};
 
 /// Common imports for users of this crate.
@@ -69,4 +70,5 @@ pub mod prelude {
     pub use crate::la_public_key::{LAPublicKey, SecKeyAlgorithm, SecKeyExchangeParameters};
     pub use crate::la_right::{LARight, LARightState};
     pub use crate::la_right_store::LARightStore;
+    pub use security::{AccessControl, AccessControlFlags, AccessControlProtection};
 }